Added WIP text, added BF function, worked on DS, and worked on DRP
Added some WIP text to some things that are still being worked on. Add a function to BasicFunctions.py to check if a process is running, will be for the Discord Rich Presence. Worked more on the Data System, changing how the default structure of the json on how data will be stored. Worked more on the Discord Rich Presence, verified that it is indeed updating so the number count at the front has been removed. Will start work on tracking if Portal 2 is started and how many people are in a lobby.
