+++ src/Scripts/BasicFunctions.py
@@ -1,5 +1,6 @@
 import Scripts.GlobalVariables as GVars
 import os
+import psutil
 
 ##############
 # CONVERSION #
@@ -94,4 +95,13 @@ def TryFindPortal2Path():
         except Exception as e:
             print("ERROR: " + str(e))
 
+    return False
+
+def checkIfProcessRunning(processName) -> bool:
+    for proc in psutil.process_iter():
+        try:
+            if processName.lower() in proc.name().lower():
+                return True
+        except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
+            pass
     return False
