+++ src/languages/English.json
@@ -71,13 +71,13 @@
   "game_path_error-saved": "Saved path!",
   "devmod_is_active": "Dev mode is now active!",
   "devmod_copied_from_local_repo": "Copied mod files from src/ModFiles!",
-  "testing_button": "Testing",
+  "testing_button": "Test Menu",
   "language_error0_language_update": "Language Set!",
   "language_error0_relaunch_game": "Please reboot the launcher...",
   "language_menu_launch_text": "The Launcher Config involves everything related to the launcher",
   "language_menu_portal2_text": "The Portal 2 Config involves everything related to your Portal 2 play session",
   "language_menu_players_text": "The Players Menu allows to set addition players as admins",
-  "data_menu_button": "Data",
+  "data_menu_button": "(WIP) Data",
   "data_system_state_txt": "Data System Status: ",
   "data_enabled": "Enabled",
   "data_disabled": "Disabled",
