+++ src/languages/Français.json
@@ -71,13 +71,13 @@
   "game_path_error-saved": "Chemin sauvegardé!",
   "devmod_is_active": "Dev mode est maintenant actif!",
   "devmod_copied_from_local_repo": "Fichiers du mod copiés à partir du repo local!",
-  "testing_button": "Test",
+  "testing_button": "Test Menu",
   "language_error0_language_update": "Langage Mis A Jour!",
   "language_error0_relaunch_game": "Relancez Votre Lanceur...",
   "language_menu_launch_text": "La configuration du lanceur implique tout ce qui concerne le lanceur",
   "language_menu_portal2_text": "La configuration de Portal 2 implique tout ce qui concerne votre session de jeu Portal 2",
   "language_menu_players_text": "Le menu Joueurs permet de définir des joueurs supplémentaires en tant qu'administrateurs",
-  "data_menu_button": "Données",
+  "data_menu_button": "(WIP) Données",
   "data_system_state_txt": "État du Système de Données: ",
   "data_enabled": "Activé",
   "data_disabled": "Désactivé",
